+++ src/app/core-components/import-backup-modal/import-backup-modal.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ImportBackupModalComponent } from './import-backup-modal.component';
+
+describe('ImportBackupModalComponent', () => {
+  let component: ImportBackupModalComponent;
+  let fixture: ComponentFixture<ImportBackupModalComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ImportBackupModalComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(ImportBackupModalComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
