+++ src/app/core-components/import-backup-modal/import-backup-modal.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ImportBackupModalService {
+  showImportModal: BehaviorSubject<boolean>;
+  constructor() {
+    this.showImportModal = new BehaviorSubject<boolean>(false);
+   }
+}
